+++ .github/CHANGELOG.md
@@ -250,9 +250,18 @@
 
 <h3>Improvements</h3>
 
+* Hamiltonians are now natively supported on the `default.qubit` device if `shots=None`. 
+  This makes VQE workflows a lot faster in some cases.
+  [(#1551)](https://github.com/PennyLaneAI/pennylane/pull/1551)
+
+* A gradient recipe for Hamiltonian coefficients has been added. This makes it possible 
+  to compute parameter-shift gradients of these coefficients on devices that natively 
+  support Hamiltonians.
+  [(#1551)](https://github.com/PennyLaneAI/pennylane/pull/1551)
+
 * The device test suite has been expanded to cover more qubit operations and observables.
   [(#1510)](https://github.com/PennyLaneAI/pennylane/pull/1510)
-  
+
 * The `MultiControlledX` class now inherits from `Operation` instead of `ControlledQubitUnitary` which makes the `MultiControlledX` gate a non-parameterized gate.
   [(#1557)](https://github.com/PennyLaneAI/pennylane/pull/1557)